@@ -9,20 +9,22 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "lint": "ng lint"
+    "lint": "ng lint",
+    "postinstall": "patch-package"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/cdk": "~19.2.17",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/material": "~19.2.17",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/router": "^19.0.0",
+    "@angular/animations": "^20.1.7",
+    "@angular/cdk": "~20.1.6",
+    "@angular/common": "^20.1.7",
+    "@angular/compiler": "^20.1.7",
+    "@angular/core": "^20.1.7",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "^20.1.7",
+    "@angular/material": "~20.1.6",
+    "@angular/platform-browser": "^20.1.7",
+    "@angular/platform-browser-dynamic": "^20.1.7",
+    "@angular/router": "^20.1.7",
     "@auth0/angular-jwt": "^5.2.0",
     "@capacitor/app": "7.0.1",
     "@capacitor/core": "7.2.0",
@@ -32,10 +34,11 @@
     "@ionic/angular": "^8.5.7",
     "@ionic/cli": "^7.2.1",
     "autoprefixer": "^10.4.16",
-    "chart.js": "^4.4.9",
+    "chart.js": "4.4.4",
+    "firebase": "^11.10.0",
     "ionicons": "^7.0.0",
     "jwt-decode": "^4.0.0",
-    "ng2-charts": "^8.0.0",
+    "ng2-charts": "8.0.0",
     "postcss": "^8.4.31",
     "rxjs": "~7.8.0",
     "tailwindcss": "^3.3.5",
@@ -43,15 +46,15 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.0",
     "@angular-eslint/builder": "^19.0.0",
     "@angular-eslint/eslint-plugin": "^19.0.0",
     "@angular-eslint/eslint-plugin-template": "^19.0.0",
     "@angular-eslint/schematics": "^19.0.0",
     "@angular-eslint/template-parser": "^19.0.0",
-    "@angular/cli": "^19.0.0",
-    "@angular/compiler-cli": "^19.0.0",
-    "@angular/language-service": "^19.0.0",
+    "@angular/build": "^20.1.6",
+    "@angular/cli": "^20.1.6",
+    "@angular/compiler-cli": "^20.1.7",
+    "@angular/language-service": "^20.1.7",
     "@capacitor/cli": "7.2.0",
     "@ionic/angular-toolkit": "^12.0.0",
     "@types/jasmine": "~5.1.0",
@@ -68,7 +71,8 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.6.3"
+    "patch-package": "^8.0.0",
+    "typescript": "~5.8.3"
   },
   "description": "An Ionic project"
 }
