@@ -12,21 +12,33 @@ import { SharedModule } from './shared/shared.module/shared.module';
 // Importar el módulo Core
 import { CoreModule } from './core/core.module';
 
+// Firebase imports
+import { provideFirebaseApp } from '@angular/fire/app';
+import { initializeApp } from 'firebase/app';
+import { environment } from '../environments/environment';
+import { provideAuth, getAuth } from '@angular/fire/auth';
+import { provideFirestore, getFirestore } from '@angular/fire/firestore';
+import { provideStorage, getStorage } from '@angular/fire/storage';
+
 /**
  * Módulo principal de la aplicación
- * 
+ *
  * Importa y configura todos los módulos necesarios para la aplicación,
  * incluyendo el módulo Core que contiene servicios e interceptores.
  */
 @NgModule({
   declarations: [AppComponent],
   imports: [
-    BrowserModule, 
+    BrowserModule,
     HttpClientModule,
-    IonicModule.forRoot(), 
+    IonicModule.forRoot(),
     AppRoutingModule,
     SharedModule,
-    CoreModule // Importar el módulo Core
+    CoreModule, // Importar el módulo Core
+    // provideFirebaseApp(() => initializeApp(environment.firebase)),
+    // provideAuth(() => getAuth()),
+    // provideFirestore(() => getFirestore()),
+    // provideStorage(() => getStorage())
   ],
   providers: [
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
