@@ -2,24 +2,24 @@ import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { ToastController, IonicModule } from '@ionic/angular';
-import { CommonModule } from '@angular/common';
+
 
 @Component({
   selector: 'app-auth',
   templateUrl: './auth.page.html',
   styleUrls: ['./auth.page.scss'],
   standalone: true,
-  imports: [CommonModule, FormsModule, ReactiveFormsModule, IonicModule]
+  imports: [FormsModule, ReactiveFormsModule, IonicModule]
 })
 export class AuthPage {
   segment = 'login';
   isLoading = false;
-  
+
   loginForm: FormGroup = this.formBuilder.group({
     email: ['', [Validators.required, Validators.email]],
     password: ['', [Validators.required, Validators.minLength(6)]]
   });
-  
+
   registerForm: FormGroup = this.formBuilder.group({
     name: ['', [Validators.required, Validators.minLength(2)]],
     email: ['', [Validators.required, Validators.email]],
@@ -51,12 +51,12 @@ export class AuthPage {
     }
 
     this.isLoading = true;
-    
+
     try {
       // Simulate API call
       await new Promise(resolve => setTimeout(resolve, 1500));
       console.log('Login data:', this.loginForm.value);
-      
+
       await this.presentToast('Inicio de sesión exitoso', 'success');
       this.router.navigate(['/dashboard']);
     } catch (error) {
@@ -74,12 +74,12 @@ export class AuthPage {
     }
 
     this.isLoading = true;
-    
+
     try {
       // Simulate API call
       await new Promise(resolve => setTimeout(resolve, 1500));
       console.log('Register data:', this.registerForm.value);
-      
+
       await this.presentToast('Registro exitoso. Bienvenido/a!', 'success');
       this.router.navigate(['/dashboard']);
     } catch (error) {
@@ -92,12 +92,12 @@ export class AuthPage {
 
   async handleSocialLogin(provider: string) {
     this.isLoading = true;
-    
+
     try {
       // Simulate API call
       await new Promise(resolve => setTimeout(resolve, 1500));
       console.log(`Login with ${provider}`);
-      
+
       await this.presentToast(`Inicio de sesión con ${provider} exitoso`, 'success');
       this.router.navigate(['/dashboard']);
     } catch (error) {
@@ -129,7 +129,7 @@ export class AuthPage {
 
   get loginEmail() { return this.loginForm.get('email'); }
   get loginPassword() { return this.loginForm.get('password'); }
-  
+
   get registerName() { return this.registerForm.get('name'); }
   get registerEmail() { return this.registerForm.get('email'); }
   get registerPassword() { return this.registerForm.get('password'); }
