@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { RouterModule } from '@angular/router';
@@ -9,7 +9,7 @@ import { RouterModule } from '@angular/router';
   templateUrl: './dashboard.page.html',
   styleUrls: ['./dashboard.page.scss'],
   standalone: true,
-  imports: [CommonModule, FormsModule, IonicModule, RouterModule]
+  imports: [FormsModule, IonicModule, RouterModule]
 })
 export class DashboardPage {
   // Datos de progreso
@@ -33,32 +33,32 @@ export class DashboardPage {
 
   // Actividades recientes
   recentActivities = [
-    { 
-      title: 'Quiz de Use of English', 
-      description: '8/10 respuestas correctas', 
-      time: 'Hace 2 horas', 
-      icon: 'list-circle' 
+    {
+      title: 'Quiz de Use of English',
+      description: '8/10 respuestas correctas',
+      time: 'Hace 2 horas',
+      icon: 'list-circle'
     },
-    { 
-      title: 'Práctica de Reading', 
-      description: 'Completado texto B2', 
-      time: 'Ayer', 
-      icon: 'book' 
+    {
+      title: 'Práctica de Reading',
+      description: 'Completado texto B2',
+      time: 'Ayer',
+      icon: 'book'
     },
-    { 
-      title: 'Test de Diagnóstico', 
-      description: 'Nivel B2 estimado', 
-      time: 'Hace 3 días', 
-      icon: 'brain' 
+    {
+      title: 'Test de Diagnóstico',
+      description: 'Nivel B2 estimado',
+      time: 'Hace 3 días',
+      icon: 'brain'
     }
   ];
 
   // Acciones rápidas
   quickActions = [
     { name: 'Iniciar Quiz Rápido', href: '/quiz', icon: 'brain' },
-    { name: 'Práctica por Tema', href: '/practice', icon: 'book' },
+    { name: 'Práctica por Tema', href: '/quiz', icon: 'book' },
     { name: 'Ver Estadísticas', href: '/stats', icon: 'stats-chart' },
-    { name: 'Plan de Estudio', href: '/calendar', icon: 'calendar' }
+    { name: 'Plan de Estudio', href: '/dashboard', icon: 'calendar' }
   ];
 
   // Pestañas de práctica recomendada
@@ -70,60 +70,60 @@ export class DashboardPage {
       {
         title: 'Comprensión de Textos Largos',
         description: 'Practica la comprensión de textos académicos y literarios extensos.',
-        href: '/practice/reading/long-texts'
+        href: '/quiz'
       },
       {
         title: 'Identificación de Ideas Principales',
         description: 'Mejora tu capacidad para identificar las ideas principales en textos complejos.',
-        href: '/practice/reading/main-ideas'
+        href: '/quiz'
       }
     ],
     listening: [
       {
         title: 'Comprensión de Conversaciones',
         description: 'Practica la comprensión de diálogos y conversaciones en diferentes contextos.',
-        href: '/practice/listening/conversations'
+        href: '/quiz'
       },
       {
         title: 'Identificación de Detalles',
         description: 'Mejora tu capacidad para captar detalles específicos en audios.',
-        href: '/practice/listening/details'
+        href: '/quiz'
       }
     ],
     useOfEnglish: [
       {
         title: 'Phrasal Verbs',
         description: 'Practica el uso correcto de phrasal verbs en diferentes contextos.',
-        href: '/practice/use-of-english/phrasal-verbs'
+        href: '/quiz'
       },
       {
         title: 'Word Formation',
         description: 'Mejora tu capacidad para formar palabras con prefijos y sufijos.',
-        href: '/practice/use-of-english/word-formation'
+        href: '/quiz'
       }
     ],
     writing: [
       {
         title: 'Essays',
         description: 'Practica la escritura de ensayos argumentativos y de opinión.',
-        href: '/practice/writing/essays'
+        href: '/quiz'
       },
       {
         title: 'Emails Formales',
         description: 'Mejora tu capacidad para escribir emails en contextos formales.',
-        href: '/practice/writing/formal-emails'
+        href: '/quiz'
       }
     ],
     speaking: [
       {
         title: 'Descripción de Imágenes',
         description: 'Practica la descripción detallada de imágenes y fotografías.',
-        href: '/practice/speaking/image-description'
+        href: '/quiz'
       },
       {
         title: 'Expresión de Opiniones',
         description: 'Mejora tu capacidad para expresar y defender opiniones en inglés.',
-        href: '/practice/speaking/opinions'
+        href: '/quiz'
       }
     ]
   };
